+++ flames/move_weights.py
@@ -13,10 +13,10 @@ class MoveWeights:
 
     insertion: float = 1.0
     deletion: float = 1.0
-    translation: float = 1.0
-    rotation: float = 1.0
+    translation: float = 0.0
+    rotation: float = 0.0
     reinsertion: float = 0.0
-    position_swap: float = 0.0
+    identity_swap: float = 0.0
     nve_md: float = 0.0
     nvt_md: float = 0.0
     npt_md: float = 0.0